Force amount to be saved as an decimal
Since our amount will be coming in as a string, I overrided the default
amount setter to make sure that the value is being converted over to a
decimal. In the case that we send it nil data or a string that is not
numeric, I created a helper function numeric? that checks that the
string is indeed numeric and can be converted over to a decimal