Add ability to create payments for a loan
This commit allows a JSON API response to create payments for a specified
loan. The testing suite only tests the happy path at the moment. The next
commits will test sad paths along with any refactoring that is needed.

To hit the endpoint you would send a `POST` request to `/api/v1/loans/:loan_id/payments`
along with the payment data. The payment data takes the signature of:
`{ "payment": { "amount" : 200.0 } }`

The response returns a 201 created status and returns the object that
was just created