Display error message if a payment exceeds outstanding balance
If there are any errors on the payment object, display the errors to the client
along with a 400 error. In the case of a payment exceeding the outstanding
balance of a loan, add a descriptive error message to the payment object
errors to be displayed along with any other validation errors.